refactor: implementa roteamento e estrutura de páginas com tela de login e dashboard, complementa o arquivo .gitignore e cria o arquivo vite.config.js
