+++ .gitignore
@@ -22,3 +22,9 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+# Variaveis de Ambiente
+.env
+.env.local
+.env.development
+.env.production
