+++ vite.config.js
@@ -1,0 +1,9 @@
+// vite.config.js
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+  server: {
+    port: 5173,
+    open: true
+  }
+});
